+++ libraries/joomla/application/web/languageprovider.php
@@ -1,0 +1,56 @@
+<?php
+/**
+ * @package     Joomla.Libraries
+ * @subpackage  Service
+ *
+ * @copyright   Copyright (C) 2005 - 2015 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+
+defined('JPATH_PLATFORM') or die;
+
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+use Joomla\Registry\Registry;
+
+/**
+ * The Web Service provider which loads the document, etc.
+ *
+ * @since  4.0
+ */
+class JApplicationWebLanguageprovider implements ServiceProviderInterface
+{
+	/**
+	 * The application object.
+	 *
+	 * @var    JApplicationWeb
+	 * @since  4.0
+	 */
+	private $app;
+
+	/**
+	 * Public constructor.
+	 *
+	 * @param   JApplicationWeb  $app  The application object.
+	 *
+	 * @since   4.0
+	 */
+	public function __construct(JApplicationWeb $app)
+	{
+		$this->app = $app;
+	}
+
+	public function register(Container $container)
+	{
+		// Setting the callables for the language
+		$container->set('language', array($this, 'getLanguage'), true, false);
+	}
+
+	public function getLanguage(Container $container)
+	{
+		$conf = $container->get('config');
+		$locale = $conf->get('language');
+		$debug = $conf->get('debug_lang');
+		return JLanguage::getInstance($locale, $debug);
+	}
+}
